backend/requirements.txt
fastapi
uvicorn
pymongo
passlib[bcrypt]
PyJWT

<<<<<<< HEAD
absl-py==2.2.0
aiofiles==24.1.0
annotated-types==0.7.0
anyio==4.9.0
astunparse==1.6.3
audioread==3.0.1
bcrypt==4.3.0
blinker==1.9.0
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
decorator==5.2.1
dnspython==2.7.0
email_validator==2.2.0
fastapi==0.115.11
ffmpeg==1.4
Flask==3.1.0
flatbuffers==25.2.10
gast==0.6.0
google-ai-generativelanguage==0.6.15
google-api-core==2.24.2
google-api-python-client==2.165.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-generativeai==0.8.4
google-pasta==0.2.0
googleapis-common-protos==1.69.2
greenlet==3.1.1
grpcio==1.71.0
grpcio-status==1.71.0
h11==0.14.0
h5py==3.13.0
httplib2==0.22.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.4.2
keras==3.9.0
lazy_loader==0.4
libclang==18.1.1
librosa==0.11.0
llvmlite==0.44.0
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
ml_dtypes==0.5.1
motor==3.7.0
msgpack==1.1.0
namex==0.0.8
numba==0.61.0
numpy==2.1.3
opt_einsum==3.4.0
optree==0.14.1
packaging==24.2
passlib==1.7.4
platformdirs==4.3.7
pooch==1.8.2
proto-plus==1.26.1
protobuf==5.29.4
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
pydub==0.25.1
Pygments==2.19.1
PyJWT==2.10.1
pymongo==4.11.3
pyparsing==3.2.1
python-multipart==0.0.20
requests==2.32.3
rich==13.9.4
rsa==4.9
scikit-learn==1.6.1
scipy==1.15.2
setuptools==77.0.3
six==1.17.0
sniffio==1.3.1
soundfile==0.13.1
soxr==0.5.0.post1
SQLAlchemy==2.0.39
starlette==0.46.1
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorflow==2.19.0
termcolor==2.5.0
threadpoolctl==3.6.0
tqdm==4.67.1
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.3.0
uvicorn==0.34.0
Werkzeug==3.1.3
wheel==0.45.1
wrapt==1.17.2
=======
absl-py
aiofiles
annotated-types
anyio
astunparse
audioread
bcrypt
blinker
cachetools
certifi
cffi
charset-normalizer
click
colorama
decorator
dnspython
email_validator
ffmpeg
Flask
flatbuffers
gast
google-ai-generativelanguage
google-api-core
google-api-python-client
google-auth
google-auth-httplib2
google-generativeai
google-pasta
googleapis-common-protos
greenlet
grpcio
grpcio-status
h11
h5py
httplib2
idna
itsdangerous
Jinja2
joblib
keras
lazy_loader
libclang
librosa
llvmlite
Markdown
markdown-it-py
MarkupSafe
mdurl
ml_dtypes
motor
msgpack
namex
numba
numpy
opt_einsum
optree
packaging
passlib
platformdirs
pooch
proto-plus
protobuf
pyasn1
pyasn1_modules
pycparser
pydantic
pydantic_core
pydub
Pygments
pymongo
pyparsing
python-multipart
requests
rich
rsa
scikit-learn
scipy
setuptools
six
sniffio
soundfile
soxr
SQLAlchemy
starlette
tensorboard
tensorboard-data-server
tensorflow
termcolor
threadpoolctl
tqdm
typing_extensions
uritemplate
urllib3
uvicorn
Werkzeug
wheel
wrapt
>>>>>>> 3eb2467 (Your commit message)
